+++ legal_knowledge_watch/services/__init__.py
@@ -7,6 +7,7 @@ from . import secrets_service
 from . import piste_oauth_client
 from . import legifrance_connector
 from . import openfisca_connector
+from . import open_lefebvre_dalloz_connector
 from . import relevance_service
 from . import storage_service
 from . import storage_dms